`hw/spinal/libcheesevoyage/gfx/lcvVgaCtrlMod.scala`: `LcvVgaCtrl`: change `clkRate` parameter to be of type `Double` (was `Int`)

+++ hw/spinal/libcheesevoyage/gfx/lcvVgaCtrlMod.scala
@@ -156,7 +156,7 @@ case class LcvVgaCtrlIo(
 }
 //--------
 case class LcvVgaCtrl(
-  clkRate: Int,
+  clkRate: Double,
   timingInfo: LcvVgaTimingInfo,
   fifoDepth: Int,
   rgbConfig: RgbConfig,
